# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ADMS HR API",
        "description": "API untuk integrasi data absensi antara ADMS dan aplikasi HR eksternal",
        "contact": {
            "email": "admin@adms.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/admin/webhooks": {
            "get": {
                "tags": [
                    "Webhook Management"
                ],
                "summary": "List all webhook configurations",
                "description": "Retrieve a list of all webhook configurations (admin only)",
                "operationId": "listWebhooks",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/WebhookConfig"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Webhook Management"
                ],
                "summary": "Create a new webhook configuration",
                "description": "Configure a new webhook endpoint to receive attendance data",
                "operationId": "createWebhook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WebhookConfigRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Webhook created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/WebhookConfig"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/webhooks/{id}": {
            "get": {
                "tags": [
                    "Webhook Management"
                ],
                "summary": "Get a webhook configuration",
                "description": "Retrieve a specific webhook configuration by ID",
                "operationId": "getWebhook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Webhook configuration ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/WebhookConfig"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Webhook not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Webhook Management"
                ],
                "summary": "Update a webhook configuration",
                "description": "Update an existing webhook configuration",
                "operationId": "updateWebhook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Webhook configuration ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WebhookConfigRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/WebhookConfig"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Webhook not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Webhook Management"
                ],
                "summary": "Delete a webhook configuration",
                "description": "Remove a webhook configuration",
                "operationId": "deleteWebhook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Webhook configuration ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Webhook configuration deleted successfully"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Webhook not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/webhooks/{id}/test": {
            "post": {
                "tags": [
                    "Webhook Management"
                ],
                "summary": "Test a webhook configuration",
                "description": "Send a test payload to verify webhook configuration",
                "operationId": "testWebhook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Webhook configuration ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "http_status_code": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "response_body": {
                                                    "type": "string",
                                                    "example": "OK"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Webhook not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Standard error response format",
                "properties": {
                    "success": {
                        "description": "Always false for error responses",
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "description": "Error details",
                        "properties": {
                            "code": {
                                "description": "Error code",
                                "type": "string",
                                "example": "INVALID_TOKEN"
                            },
                            "message": {
                                "description": "Human-readable error message",
                                "type": "string",
                                "example": "Token tidak valid atau expired"
                            },
                            "details": {
                                "description": "Additional error context",
                                "type": "object",
                                "example": []
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "WebhookConfig": {
                "title": "Webhook Configuration",
                "description": "Webhook configuration model",
                "properties": {
                    "id": {
                        "description": "Webhook configuration ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Webhook name/description",
                        "type": "string",
                        "example": "HR System Webhook"
                    },
                    "url": {
                        "description": "Webhook URL (must be HTTPS)",
                        "type": "string",
                        "format": "uri",
                        "example": "https://hr.example.com/api/webhooks/attendance"
                    },
                    "is_active": {
                        "description": "Whether the webhook is active",
                        "type": "boolean",
                        "example": true
                    },
                    "secret_key": {
                        "description": "Secret key for signature verification",
                        "type": "string",
                        "example": "your-secret-key-here"
                    },
                    "headers": {
                        "description": "Additional HTTP headers to send",
                        "type": "object",
                        "example": {
                            "X-Custom-Header": "value"
                        },
                        "nullable": true
                    },
                    "retry_attempts": {
                        "description": "Number of retry attempts on failure",
                        "type": "integer",
                        "example": 3
                    },
                    "timeout": {
                        "description": "Request timeout in seconds",
                        "type": "integer",
                        "example": 30
                    },
                    "created_at": {
                        "description": "Configuration creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-15T08:00:00Z"
                    },
                    "updated_at": {
                        "description": "Configuration last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-15T08:00:00Z"
                    }
                },
                "type": "object"
            },
            "WebhookConfigRequest": {
                "title": "Webhook Configuration Request",
                "description": "Request body for creating/updating webhook configuration",
                "required": [
                    "name",
                    "url",
                    "secret_key"
                ],
                "properties": {
                    "name": {
                        "description": "Webhook name/description",
                        "type": "string",
                        "example": "HR System Webhook"
                    },
                    "url": {
                        "description": "Webhook URL (must be HTTPS)",
                        "type": "string",
                        "format": "uri",
                        "example": "https://hr.example.com/api/webhooks/attendance"
                    },
                    "is_active": {
                        "description": "Whether the webhook is active",
                        "type": "boolean",
                        "example": true
                    },
                    "secret_key": {
                        "description": "Secret key for signature verification",
                        "type": "string",
                        "example": "your-secret-key-here"
                    },
                    "headers": {
                        "description": "Additional HTTP headers to send",
                        "type": "object",
                        "example": {
                            "X-Custom-Header": "value"
                        },
                        "nullable": true
                    },
                    "retry_attempts": {
                        "description": "Number of retry attempts on failure (default: 3)",
                        "type": "integer",
                        "example": 3
                    },
                    "timeout": {
                        "description": "Request timeout in seconds (default: 30)",
                        "type": "integer",
                        "example": 30
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}